requirements.txt
Django~=3.2.12
kafka-python~=2.0.2
djongo ~=1.3.6